requirements.txt
streamlit
sentence-transformers
faiss-cpu
Pymupdf
numpy
